requirements.txt
streamlit
pandas
numpy
torch
clip-openai
faiss-cpu
pillow
tqdm